pyproject.toml
[project]
name = "areteusml"
version = "0.1.0"
description = "Production ML text classification pipeline for Banking77 intent detection"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11,<3.13"
authors = [{ name = "Daniel Alexis Cruz", email = "danielalexiscruz.pro@gmail.com" }]

dependencies = [
    # ML Core
    "torch>=2.6.0,<3",
    "transformers==4.49.0",
    "datasets==3.3.2",
    "tokenizers==0.21.1",
    "accelerate==1.4.0",
    "setfit==1.1.1",
    "sentence-transformers==4.0.1",
    "scikit-learn==1.6.1",
    "optuna==4.3.0",
    "shap==0.46.0",
    # Data
    "pandas==2.2.3",
    "numpy==1.26.4",
    "pandera==0.21.1",
    "dvc>=3.59,<4",
    "dvc-s3>=3.2,<4",
    "pathspec<1",
    # Experiment Tracking
    "mlflow==2.21.0",
    # Inference
    "onnx==1.17.0",
    "onnxruntime==1.21.0",
    "optimum==1.25.0",
    # API
    "fastapi==0.115.11",
    "uvicorn[standard]==0.34.0",
    "pydantic==2.10.6",
    "pydantic-settings==2.8.1",
    "python-jose[cryptography]==3.4.0",
    "passlib[bcrypt]==1.7.4",
    "redis==5.2.1",
    "slowapi==0.1.9",
    "httpx==0.28.1",
    # Monitoring
    "evidently==0.6.7",
    "prometheus-client==0.22.0",
    # Dashboard
    "streamlit==1.43.2",
    "streamlit-shadcn-ui==0.1.19",
    "streamlit-elements==0.1.0",
    "plotly>=5.10.0,<6",
    # Orchestration
    "dagster==1.10.10",
    "dagster-webserver==1.10.10",
    # Database
    "sqlalchemy==2.0.38",
    "psycopg2-binary==2.9.10",
    "alembic==1.14.1",
    # Utilities
    "python-dotenv==1.0.1",
    "loguru==0.7.3",
    "rich==13.9.4",
    "typer==0.15.2",
    "onnxscript>=0.6.2",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.5",
    "pytest-asyncio==0.25.3",
    "pytest-cov==6.0.0",
    "ruff==0.11.0",
    "pre-commit==4.1.0",
    "locust==2.33.0",
    "ipykernel==6.29.5",
    "jupyter==1.1.1",
    "matplotlib==3.10.1",
    "seaborn==0.13.2",
]

labeling = [
    "argilla==2.6.0",
]

[tool.uv]
environments = ["sys_platform == 'win32'"]

[[tool.uv.index]]
name = "pytorch-cu126"
url = "https://download.pytorch.org/whl/cu126"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu126" }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["ml", "backend"]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "S",   # flake8-bandit
    "T20", # flake8-print
    "SIM", # flake8-simplify
    "TCH", # flake8-type-checking
    "RUF", # ruff-specific
]
ignore = [
    "S101",  # assert used (needed for tests)
    "S105",  # hardcoded password (false positives in config)
    "E501",  # line too long (handled by formatter)
    "B008",  # Depends() in defaults (FastAPI pattern)
    "N803",  # X_train/X_test naming (ML convention)
    "N806",  # X_train/X_test variables (ML convention)
    "S104",  # binding to 0.0.0.0 (intentional for Docker)
    "S301",  # pickle load (trusted internal models only)
    "TC002", # type-checking imports (over-optimization)
]

[tool.ruff.lint.per-file-ignores]
"ml/notebooks/*" = ["T20", "E402"]
"backend/tests/*" = ["S101", "S106"]
"ml/tests/*" = ["S101"]
"ml/utils/reproducibility.py" = ["T20"]

[tool.pytest.ini_options]
testpaths = ["backend/tests", "ml/tests"]
asyncio_mode = "auto"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "gpu: marks tests requiring GPU",
    "integration: marks integration tests",
]
